Add experimental NewPlayer class and NewDeck subclass
Rename NewCards class to NewCard

Cards are stored within objects (or "Players") created with the
NewPlayer class. drawOn() draws all the stored cards on the
screen and hitbox() is an integer that returns the pressed card
based on mouse position.

NewDeck objects (or "Decks") behave a bit differently; hitbox()
is a boolean that returns true only if the last card was
pressed, drawOn places Cards in a way that resembles a card
deck and a constructor that builds the entire card set.

eraseCard() and insertCard() allows "Players" and "Decks" to
pass cards with each other.